декомпозиция редьюсера
