+++ package.json
@@ -33,6 +33,7 @@
     "node": "10"
   },
   "dependencies": {
+    "axios": "0.19.2",
     "leaflet": "1.6.0",
     "moment": "2.24.0",
     "prop-types": "15.7.2",
@@ -40,6 +41,8 @@
     "react-dom": "16.12.0",
     "react-redux": "7.2.0",
     "react-router-dom": "5.1.2",
-    "redux": "4.0.5"
+    "redux": "4.0.5",
+    "redux-thunk": "2.3.0",
+    "reselect": "4.0.0"
   }
 }
