+++ src/components/app/app.jsx
@@ -2,11 +2,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Main from '../main/main.jsx';
 import {Switch, Route, BrowserRouter} from 'react-router-dom';
-
+import {ActionCreator} from '../../reducer/data/data.js'
 import AboutOffer from '../about-offer/about-offer.jsx';
 import {AppRoute} from '../../const.js';
 import {connect} from 'react-redux';
-
+import NameSpace from '../../reducer/name-space.js';
+import combineReducers from '../../reducer/reducer.js';
 const App = (props) => {
   const {offerCards, cities, activeCity} = props;
 
@@ -60,12 +61,14 @@ App.propTypes = {
   }).isRequired,
 };
 
-const mapStateToProps = (state) => (
-  {
-    activeCity: state.city,
-    offerCards: state.offers
+const mapStateToProps = (state) => {
+  console.log(state.DATA);
+  return ({
+    activeCity: state.DATA.city,
+    offerCards: state.DATA.offers
   }
-);
+  )
+};
 
 export {App};
 export default connect(mapStateToProps)(App);
